+++ src/engine/inject/references.js
@@ -1,20 +1,16 @@
-var referenceRegex = require('../regex/reference')
-var walkGlobal = require('../walkGlobal')
+  var referenceRegex = require('../regex/reference')
+  var walkGlobal = require('../walkGlobal')
 
-/**
- * Inject references to functions.
- *
- * @api private
- *
- * @param {Object} funcs reference
- * @param {Object} task
- */
+  /**
+  * Inject references to functions.
+  *
+  * @param {Object} funcs reference
+  * @param {Object} task
+  */
 
 function injectReferences (funcs, task) {
   /**
    * Inject task.
-   *
-   * @api private
    */
 
   function inject (taskKey) {
@@ -24,8 +20,6 @@ function injectReferences (funcs, task) {
 
     /**
      * Inject reference.
-     *
-     * @api private
      */
 
     function reference () {