+++ src/engine/inject/strings.js
@@ -3,8 +3,6 @@ var quotedRegex = require('../regex/quoted')
 /**
  * Inject functions that return strings.
  *
- * @api private
- *
  * @param {Object} funcs reference
  * @param {Object} task collection
  */
@@ -12,8 +10,6 @@ var quotedRegex = require('../regex/quoted')
 function injectStrings (funcs, task) {
   /**
    * Inject a function that returns a string.
-   *
-   * @api private
    */
 
   function inject (taskKey) {